// libs/store/Review/actions-type.tsx
export enum ProductReviewActionType {
    ADDING_PRODUCT_REVIEW = 'catalog/ADDING_PRODUCT_REVIEW',
    ADD_PRODUCT_REVIEW_SUCCESS = 'catalog/ADD_PRODUCT_REVIEW_SUCCESS',
    ADD_PRODUCT_REVIEW_FAILED = 'catalog/ADD_PRODUCT_FAILED ',
    GET_ALL_PRODUCT_REVIEWS = 'catalog/GET_ALL_PRODUCT_REVIEWS',
    GET_ALL_PRODUCT_REVIEWS_SUCCESS = 'catalog/GET_ALL_PRODUCT_REVIEWS_SUCCESS',
    GET_ALL_PRODUCT_REVIEWS_FAILED = 'catalog/GET_ALL_PRODUCT_REVIEWS_FAILED',
}